requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.1
django-seed==0.3.1
djangorestframework==3.14.0
Faker==20.1.0
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
toposort==1.10
djangorestframework-simplejwt==5.3.1.